Store uploaded file, add validation logic
This stores an uploaded file to `./temp` when it is `post`ed to
the same endpoint as the editing page.

It treats the process of uploading a file as a _validation_, much
like checking if a user has inputted service title under 100
characters, or a percentage as just a number.

If the validation fails (eg no file is provided) then it passes a
message to the template on the `errors` object.

This is built in such a way that a dictionary of validations
is provided in the content, eg:
```yaml
validations
  -
    name: is_acceptable_file_size
    message: Your file is too big
  -
    name: is_open_document
    message: That's not an open document format
```
…and then it's up to the app to implement a function called
`is_open_document` which returns `True` or `False` when provided
`post`ed data.