Disable authentication when testing the app
This commit configures the app not to authenticate in a testing
environment. This means that we can test real features of the app
going forward, rather than just testing if we get a 401 response.

To be consistent with our naming scheme the environment (test,
development, production) is set with an environment variable
called `DM_ADMIN_FRONTEND_ENVIRONTMENT`