Add a login page
This commit replaces the basic auth with a login page and cookie-based sessions.

It does not do any user management, but does use Flask's session handling.
Session lifetime is 1 hour, and is renewed on every request.

It adds a new environment variable (`DM_ADMIN_FRONTEND_COOKIE_SECRET`) which
is used to encrypt the cookie. This will need to be set to something secret
when deploying the app.