Add timestamp to uploaded file instead of the old version
New files are created with 'YYY-MM-HH-hhmm' filename suffix.
Since file names are now unique there's no need to rename old
document versions.

Services are archived with URLs of the previous document version
and the service is then updated with the new file names. That way the
archived services are always linked with the documents that were
available at the time.