Fix bug with pricing list
There was a bug when max price was left empty. It meant that max price got
replaced with the value of the next item in the pricing list (unit). This is
because any list submitted to the app was having empty items stripped out.

This commit changes the app to only strip empty items from lists that get
checked for their length.