Deny third-party sites from embedding in an iframe
Added an `X-Frame-Options` line in our header which denies
embedding our site in an iframe by other sites.
All very well and good for this app, but we'll also have to
duplicate the code in other apps we want to share this behaviour.
Things to consider:
- We *could* set this at the server level if we wanted to make
this change more uniform.
- The alphagov/whitehall team have set their public-facing headers
for their content to `ALLOWALL`, but @minglis says we're not going
to allow this at present.
https://github.com/alphagov/whitehall/commit/9f8c7ccd1fdf7695e66b1340cb6b89ca0e134f9a