Frameworks content based document uploading
This change uses the same method of editing services as the
supplier frontend. It also fixes the tests and updates how the
configuration is managed.

Config:
Standardise on DM_ prefixed names and remove `os.getenv` calls (as this
is handled for us by dmutils.config).

Tests:
A few of the tests were failing because the stub service data didn't
have a lot which the frameworks content depends on.