Add ability to edit supplier declarations
This adds the ability for a CCS Sourcing team member to view and edit
G-Cloud 7 supplier declarations. In the supplier listing the only link
they will see alongside a supplier is 'G-Cloud 7 declaration'. This will
take them to the G-Cloud 7 supplier declaration overview where they can
edit each section. This uses the new ContentLoader but hard codes the
G-Cloud 7 link in the supplier list as it is the only framework they can
edit at the moment.

Note that this does no validation, it is assumed that the CCS Sourcing
team know the rules.