Multi-extension agreement download
This change adds support for framework agreements potentially having
different extensions introduced in
https://github.com/alphagov/digitalmarketplace-supplier-frontend/pull/340

Because the full path is no longer known, instead the prefix is provided
and the most recent matching file is returned. This works because the
agreement file names do not overlap and the timestamp is added to the
start of the archive file name. If the supplier first uploads a PDF and
then uploads a JPEG the JPEG will be returned here because it is the
most recent matching upload.