Upgrade frameworks and toolkit
This brings in the latest frontend toolkit and frameworks to make the
pricing work again. This was broken by an earlier update to the utils
package. This is a massive jump in versions but the only place that
should be affected is the service editing and that has been tested.