Page for changes to service status
This is kinda hacky but I want to see if it’s the right approach before cleaning
it up.

Shows a page of audit events for change of service status. One page per day.
Pagination to move to the next/previous days that have events.

I think this is easier to understand than the weird date input filter thing.