Make pagination work for days with lots of events
If a day had lots of events they would need to be shown across multiple pages
(because the API, currently, only returns 100 events).

This modifies the pagination to work for both days and pages within days. So
the journey now looks like:

Wednesday → Thursday → Thursday page 2  → Thursday page 3 → Friday

It does mean that you can’t skip days with no events when paginating, but this
would have fallen apart on days with many events in the previous version
anyway.