Remove obsolete `page_heading` macro
The `page_heading` macro has been replaced by the `page-heading` element
in the toolkit.
Only a couple of templates were using the antiquated one, so I swapped them
over to using the toolkit and removed the `macros/page-headings.html` file.

Took a bit of a shortcut sticking a `<br>` tag in the revisions template, but w/e.
It was easy, it works and no one's using it right now.