Allow admin to move users between supplier accounts
Currently the "Remove user from supplier" button doesn't work, because before we had actual buyer accounts users were ported to a "buyer" role, which now fails unless the user happened to have an email address that is allowed for our buyer acounts, and anyway is not the expected behaviour any more.

This removes the broken "Remove from supplier" button, and instead adds a form field that allows an admin user to move existing supplier users from one supplier to another.