Add fix for chrome bug with font-size-adjust
Chrome seems to be increasing the font slightly
when `font-size-adjust` is set (which we have been
doing since this change to the govuk frontend
toollkit:

alphagov/govuk_frontend_toolkit#271

The grid used for `.footer-meta` sets its columns
to `display: inline-block' (to have more control
over the vertical spacing) which can mean their
horizontal spacing is effected by changes to how
lines of content are rendered. It seems having
`font-size-adjust` set is increasing the
horizontal spacing of the line on which they sit,
meaning they can no longer git on it.