Use content loader summary to correlate questions and answers
The registered address wasn't being displayed on the summary page because the multiquestion macro needs the `.summary` method to have been called to tie questions and values together for the component parts of the multiquestion.

Once that has been done we can just pass item.name and item.value through to the relevant macro, rather than passing values from the declaration object directly.