Format rows for csv before giving to iter__csv
The iter_csv method was handling formatting the rows into lists from a
users dict, and adding the headings. This was fine until the buyer
frontend needed to share the iter_csv function as it doesn't have any
headings. This makes is neccessary to feed iter_csv the correctly
formated rows.