Update to use saved agreement paths from FrameworkAgreement
This updates the app to use the new FrameworkAgreement data to ascertain whether
an agreement has been returned and/or countersigned, rather than checking for
the existence of files in S3 and their relative timestamps.

A welcom side-effect of this work is that we can now hide links on the /suppliers
page when a signed agreement document doesn't exist, which should make life easier
for admin users trying to find documents on that page.