Add post routes to approve/on-hold agreements
When the admin user hits either the 'on hold' or 'countersign' buttons,
these routes are used to hit the api and update the framework_agreement.

The name of the organisation is also taken from the supplier_framework
variable passed into the frontend template that jinja uses. This name is
used to populate the flash message that is displayed to the user on the
next page.

STILL TO DO - The redirect on these routes is currently hardwired to teh
framework agreements list page. It needs to redirect to the next
agreement to be processed. This is out of the scope of this story
however.

Finally, admin has been added to the roles authorised to view the
`view_signed_agreements` view. Logic is used in the template on the
frontend to prevent admins from actually changing anything however.