Create buttons to put agreements on hold or approve
The buttons are linked to the post routes just created. They pass the
agreement id as well as the organisation name.

Checks are made on the logged in users role to decided if they should
see the buttons or not. If they are ccs-admin they will see them, and
the buttons are conditionally show depending on the status of the
agreement.

Also, a flash message is displayed in a notification banner if an
agreement was just put on hold/approved.