Stop using the `search-summary` pattern for our search summary
While this is admittedly a retrograde move, I think it's fine
right now.

- the buyer app is using two different search summary templates,
  none of which are from the toolkit
- this app is effectively doing its own search summary and them
  passing it into a 3-line template
- the template has nothing in it

It seems like the real thing to do if we think this is a pattern
(I think it is) is to rewrite it so that it works for all cases
and doesn't just echo an unescaped message.

This achieves the short-term goal of removing the `safe` filters
from the front-end toolkit, so that seems worth doing first.