Remove conditional links from supplier listing page
Links are currently shown conditionally depending on the existence of a file path in the supplier data.
Getting this data requires too many API calls that is making the page time out, so we need to revert
to the previous behaviour of showing all links for all suppliers without checking if the file actually exists.
(This only affects old agreements flow - i.e. G7 and DOS.)

This also adds a "jump" route for downloading agreement files, where the actual path in S3 is retrieved
before being passed through to the existing `download_agreement_file()` route.