Avoid making multiple API calls per supplier for suppliers page
The suppliers listing page was recently changed to fetch the SupplierFramework record for every supplier listed on the page.  This was in order to get the `signed_agreement_path` if it exists.  This was a nice feature on the page, but results in page timeouts and massive overload on the API, so we're reverting to the old behaviour of showing links for all suppliers without checking if the file exists or not.