Add DOS2 and remove G-Cloud 7 from ccs-sourcing view
CCS-Sourcing need to be able to amend DOS2 declarations on the behalf of suppliers.  This adds the DOS2 framework as a column on the CCS-Sourcing suppliers view.

The columns on that page become too squashed with four frameworks, so this also removes the G-Cloud 7 column of links, as agreements for that framework should no longer require work from CCS sourcing.