Remove safe escape and redo table
Remove jinja safe escape filter and line breaks.  This filter shouldn't be being used anywhere anymore.

Redo table. Pull in new version of the frotend toolkit to use rowspan on column names so that the table stucture is more sound.  Remove divs, as these shouldn't be used together with tables.