Fix bug where archived service timestamps can be equal
for the first ever edit of a service the commit_and_archive_service function in the API will create two archived services at the same time - one the original version and one the updated version.  The >= condition here was causing the page to fail when viewing the first edit of a service.