Fix missing DM_HTTP_PROTO in staging config
Staging was sending user invite emails with `http://` because staging
config didn't set the correct DM_HTTP_PROTO.

The config was copied from Live in #87 to disable CSRF checks in
staging. We couldn't find a reason why we'd still want this, so
I'm reverting the setting to match the Live config.