service updates view: redesign to perform at most two find_audit_events calls
those being the most recent unacknowledged update events and the least
recent unacknowledged update events, which gives us most of the required
information. it does leave us with an information gap if we have more
than two pages worth of events, in this case we have to settle for not
giving a proper count of the actual number of events and not being able
to say anything certain about the number of editing users.