Remove link to (broken) list of approved changes page
The list of approved changes doesn't really work at the moment, so we don't want to link to it from the home page.

I've also moved the link to "Check edits to services" above the "View statistics" link as it will be more relevant day-to-day.  And removed the unnecessary "body" text that just echoed both the heading and link text in slightly different words.