Fix missing frameworks comms timestamps after dmutils.S3 upgrade
New dmutils.S3 version requires `load_timestamp` argument to be
set in order to display the `last_modified` timestamp.

The change explicitly loads the metadata headers for the timestamps,
which means it'll display the header value if available.

Both old and new versions seem to load the oldest file date (instead
of the newest one), so the date doesn't reflect the actual last
upload time. I'm not sure if that's a bug or intended behaviour, but
I'm keeping it as is.