Don't use flash messages if not redirecting
We generally use flash messages for banners at the top of the page, but in this case we're not redirecting on error, but reloading the form.  In this case we should be passing errors to the template as errors on the form field, rather than as a flash message in the session.

The design is that we will display validation errors on the form as banner messages, so this adds a section to the template to display these, while removing the error validation wrapper from the field itself to avoid duplication.

Also updates error message content slightly based on discussion with Constance.