Pull in Javascript to clear local anchor from links following error
Fixes this issue found in QA:
If a page has two separate sections that both have a validation error. User clicks on the link at the top to take them to the section that is in error, corrects the error then attempts save the change, instead of being returned to the top of the page to show the user what is still wrong on the page, the page remains where it is.