Use toolkit styling for pricing questions rather than custom
There is no reason for us to use custom styling for pricing
questions. Instead we can use the toolkit styling.

Also note, I feel it is OK to remove the `visually-hidden` class
at the bottom of the _pricing.scss file because this is provided
by the GOV.UK application.css which is pulled in by the GOV.UK
base page which we build on top of.