BUGFIX: show most recent upload date on framework files
The "Last modified" date shown to admin users on the "upload communications and clarifications" page should be the latest date a file was uploaded, but was actually displaying the date of the first upload to the relevant bucket.

I suspect that maybe our S3 list() used to return things in reverse date order and got swapped to increasing date order at some point.  But who knows, maybe this never worked...

I've also added nicer "empty" messages for when no files are uploaded yet - it used to just say "Last modified None".