Public stats page
 ## Summary
Remove the role check on the framework application stats page. This will
allow us to display these stats e.g. in the office without having to
leave a privileged admin user logged into the marketplace. The admin
site is protected by an IP whitelist by the digitalmarketplace-router,
which should provide sufficient protection - added to the fact that the
data we show here should not be particularly sensitive and much of it is
available on the performance platform already.

 ## Ticket
https://trello.com/c/zTkPCGSY/289-please-create-an-admin-account-which-can-access-framework-application-stats-so-we-can-look-at-them-on-the-big-tv