Rejig view suppliers agreements list
The list was laid out so that new frameworks would appear in a new
column. It was getting crowded.

This switches the view so a new framework adds a new row instead.

It also pulls this list of frameworks to display from the api rather
than being a hard coded list. As we don't display all frameworks, and
some frameworks are displayed differently (old signing flow frameworks),
we need a couple of constants to control this behaviour.