Replace <supplier_id> with <int:supplier_id>
A bit of housekeeping here. The suppliers routes were a mix of <supplier_id>, which seemed to have been interepreted as a String, and <int:supplier_id> which is an Int. The mix made it hard to search the file and was just inconsistent. I've updated the tests to reflect that these endpoints will take ints