Update test to include all framework statuses
Update route to pass appropriate frameworks to view

The route now passes the right frameworks through, dependent on the user role.

Change link to ‘Download’ for old frameworks

The old framework agreements don’t work in embeds like the new ones, so the only action users can take is to download them. This is now implemented