content loader: make thread-safe through use of context-local deep copies
this implementation is based on the one in the buyer frontend and is a
little more involved than that in the the supplier frontend. this is
required because these frontends need information from the application
object and the api before they can populate their content loaders. hence
this is performed quite late in the startup process. this should all be
fine as long as no code needs to access the methods of the content loader
before the application is initialized, though in such a case the attempt
is made to fail obviously and loudly.