Prevent admins accidentally introducting whitespace
We've found five examples of admins accidentally introducing whitespace to company names. When a leading whitespace is added to a company's name, they're moved to the numerical section of the A-Z. This is not optimal for being found! I've now implemented a small tweak to make sure any accidental spaces are stripped out prior to being sent to the API.

I've added a test that fails, proving where the issue exists