Replacing toolkit/breadcrumbs with govukBreadcrumbs
Biggest thing to look out for as part of code review is that the label
and href match what was previously used on the page.

breadcrumbs block will not be available once we migrate from govuk_template
to govuk-frontend template. We will need to move these inside a beforeContent block

See:
- https://design-system.service.gov.uk/styles/layout/#limiting-width-of-content
- https://design-system.service.gov.uk/styles/page-template/