Replace toolkit/page-headings with govuk-frontend headings
Generally, the headings were `h1` with a css class of `govuk-heading-xl`.
These were the mappings I used while replacing

Generally  use `govuk-heading-xl` class
If "smaller : true" when calling toolkit/page-heading, is 'h1` with
`govuk-heading-l`.
If "Context" used when calling toolkit/page-heading, add span above the
`h1` and use govuk-caption-xl or govuk-caption-l depending on the heading
used.

while replacing the headings I noticed that there is an inconsistency with
the markup where sometimes the headings are placed inside a grid and others
arent. I also found a difference between heading titles font sizes, it
would be good to use a consistant size. I will try address these when
the grid system is replaced.