Refactor list of links
As the links on the home page, are categorised. it feels more semantic
to declare them in an unordered list. This will help screenreaders understand
how the links are relatred. I.e they are not just a group of random links
with no context.

As part of this commit, I also tidied up the indenting of the code.