Use Phasebanner instead of user research sign up banner
While we were trying to migrate from govuk_template to govuk-frontend
template. An issue was identified with having the user research sign up
banner positioned between the top black header and the next red horizontal
line. It would mean that we could not use GOV.UK Frontend header component
and we would still have to maintain the user research banner and it also
relied on govuk_templates cookie implementation which we would also need
to either copy or remove completely. As this was the only component using
GOVUK.cookie method it would be good to remove it.

This change also brings our designs closer inline with GOV.UK Design system

Stefan Nicolaou, Kat Stevens and I, meet and agreed that we would drop
the user research banner and instead take all users to the "help" page.
On the "Help" page we would add some copy for users to sign up to user
research. It will still require users to have an existing account before
they can sign up for user research.