Remove report-a-problem.js and css
My understanding is that we do not allow users to report problems as in
the past the feedback was never monitored. HTML for the component is not
included in the base page/ template layout, so it seems as though this
was accidentally left in. Therefore the js code was being bundled into
application.js and application.css and sent to the browser for no reason.