setup GOV.UK Frontend header & main content
We decided that for admin frontend we would still check the red bottom
border of the header. Another change is that the crown, gov.uk and Digital
Marketplace header text is now a single link back to the services home
page instead of gov.uk

We can now finally remove the temporary styling hacks we needed
while we were rolling out components.

Now using the new dmHeader and dmFooter components for the first time.