Splits index.html into three smaller partials
This page is overally compilicated to read and to format. I have decided
to break it out initially as three seperate partials based on the
three main conditions. This might change in the future and could
be split into 6 high level role based partials but I wanted to start off
small.