apply generic styles that were removed with govuk-template
When we removed govuk-template it was thought that it would only impact
fonts, headers and footers. It turned out that govuk-template was also
setting up default styles for things like links and paragraphs etc.
so any links that did not have a class of govuk-link were displayed
with the browser default link colours.

We were able to fix some of the issues related to missing links/paragraph
paddings by using govuk-frontend's $govuk-global-styles mixin. However,
it still didn't fix issues with "hidden" styles that hide content from
screenreaders and visually. To over come that we reintroduced some of
a slimmed down version of govuk-templates css file (see separate comitts)