Fix govuk-frontend images path
We wanted GOV.UK Frontend image assets to be in a separate folder in our
static assets directory structure; however, this doesn't fit with the
[page template for GOV.UK Frontend][1], which expects the favicon (and
other image assets) to be located in `/static/images`.

This commit changes the gulp task and Sass configuration to put GOV.UK
Frontend image assets in this location, fixing a bug with the favicon.

[1]: https://github.com/alphagov/govuk-frontend/blob/v2.13.0/src/template.njk#L17