Stop filtering re-used manifest in-place
In 58b1af0a33ac560e1798aaa64f5943bddc510d3b, we started filtering this manifest in-place for performance reasons. However, this manifest gets re-used for each draft service. This caused problems for g-cloud because there are some questions that only appear in certain lots. This meant that we would filter these questions out, and then throw a KeyError when trying to get the summary of a later draft for a different lot.

Switch back to not filtering/summarising in place to avoid this problem. This will reduce performance slightly. However, it's the admin app, so it doesn't matter that much.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>