Show last two expired G-Cloud frameworks
Admins still need to handle requests for old G-Cloud frameworks. Make this easier for them by including links for the last two expired G-Cloud frameworks on the homepage.

There was a deliberate decision to exclude old G-Cloud frameworks made in 2017. However, I can find nothing that says why the change was made. So I think we're probably safe enough to reverse that decision, given that CCS reported the current behaviour as a bug.