Add form to change framework status
This is the second part of a firebreak project to give admins in non-prod environments the ability to view and change the framework state. Follows on from https://github.com/alphagov/digitalmarketplace-admin-frontend/pull/767.

Create the form that admins will use to change the state. Wire it up so that it shows the current state. The form doesn't yet do anything on POST.

Since this is a firebreak experiment, I'm hiding this in production.