Let non-prod admins change framework statuses
This is the third part of a firebreak project to give admins in non-prod environments the ability to view and change the framework state. It follows on from https://github.com/alphagov/digitalmarketplace-admin-frontend/pull/768.

Update the 'change framework status' page to actually change the framework status. The full user journey now works.

Since this is a firebreak experiment, I'm hiding this in production.