Use govukTable for viewing users
We were previously using toolkit/summary-table.html. We want to replace all uses of the toolkit with design system components. This is part of our migration to Design System 3 to improve accessibility. Admin-frontend is not a priority app, but I was working in the area for my firebreak project, so did this in passing.

No obvious changes in behaviour - just minor formattting differences.