Let admins change user's names themselves
We occassionally get requests from users to change their name. CCS support forward these requests on to us. We change the name via the API. We don't actually provide any value - if an admin asks us to change a user's name, we do it.

So it is safe to cut out the developer in the middle and give admins the ability to change user's names themselves. This should save both sides time.

This action is already auditted via the API, so we will have a record of any changes admins make.