Allow category admins to download service edit reports
We've currently got a manual process to create the report and email it to a particular category admin once a month. I'm starting to get bored of this, so would like to automate it.

My plan is to run the report-generating script on Jenkins. Then it can upload the report to the reports bucket. Category admins will then be able to log in to the DMP to download the report. The next step will be to modify the script to upload the report to S3 and also send an email via Notify that contains a link to this new endpoint.

Category admins can approve service edits, so it makes sense to allow them to view the report on their own activities.